feat: main page, not found, app , add routes
