+++ src/components/app/app.jsx
@@ -1,25 +1,24 @@
 import React from 'react';
+import { Routes, Route } from 'react-router-dom';
 
 import styles from './app.module.css';
 
 import AppHeader from '../app-header/app-header';
-import BurgerConstructor from '../burger-constructor/burger-constructor';
-import BurgerIngredients from '../burger-ingredients/burger-ingredients';
 
-import { DndProvider } from 'react-dnd';
-import { HTML5Backend } from 'react-dnd-html5-backend';
+import MainPage from '../../pages/main-page/main-page';
+import { NotFound404 } from '../../pages/not-found-404/not-found-404';
 
 function App() {
 
     return (
         <>
             <AppHeader />
-            <main className={styles.main}>
-                <DndProvider backend={HTML5Backend}>
-                    <BurgerIngredients />
-                    <BurgerConstructor />
-                </DndProvider>
-            </main>
+
+            <Routes>
+                <Route path="/" element={<MainPage />} />
+
+                <Route path="*" element={<NotFound404 />} />
+            </Routes>
         </>
     );
 }
