+++ src/index.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import { BrowserRouter as Router } from 'react-router-dom';
+
 import { Provider } from "react-redux";
 import { configureStore } from '@reduxjs/toolkit'
 
@@ -8,9 +10,6 @@ import './index.css';
 import App from './components/app/app';
 import { rootReducer } from './services/reducers/rootreducer';
 
-import { enhancer } from './utils/utils';
-
-//const store = createStore(rootReducer, enhancer);
 //в configureStore devTools дефолтно true
 const store = configureStore({
   reducer: rootReducer
@@ -19,11 +18,13 @@ const store = configureStore({
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
-
+//BrowserRouter должен быть выше React.StrictMode иначе могут появиться ошибки
 root.render(
   <React.StrictMode>
     <Provider store={store}>
-      <App />
+      <Router>
+        <App />
+      </Router>
     </Provider>
   </React.StrictMode>
 );
