+++ src/pages/main-page/main-page.jsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { DndProvider } from 'react-dnd';
+import { HTML5Backend } from 'react-dnd-html5-backend';
+
+import styles from './main-page.module.css';
+
+import BurgerConstructor from '../../components/burger-constructor/burger-constructor';
+import BurgerIngredients from '../../components/burger-ingredients/burger-ingredients';
+
+function MainPage() {
+
+    return (
+        <>         
+            <main className={styles.main}>
+                <DndProvider backend={HTML5Backend}>
+                    <BurgerIngredients />
+                    <BurgerConstructor />
+                </DndProvider>
+            </main>
+        </>
+    );
+}
+
+export default MainPage;
