+++ src/pages/not-found-404/not-found-404.module.css
@@ -7,5 +7,14 @@
     display: flex; 
     flex-direction: column; 
     justify-content: center;
-    /*align-items: center;*/
+    margin: 0 auto;
+    align-items: center;
+}
+
+.link {
+    color: white;
+}
+
+.link:hover{
+    text-shadow: 0px 4px 32px rgba(51, 51, 255, 0.50), 0px 0px 8px rgba(51, 51, 255, 0.25), 0px 0px 16px rgba(51, 51, 255, 0.25);
 }
